+++ projects/swimlane/ngx-charts/src/lib/number-card/card.component.ts
@@ -10,13 +10,18 @@ import {
   ChangeDetectionStrategy,
   ChangeDetectorRef,
   NgZone,
-  OnDestroy
+  OnDestroy,
+  PLATFORM_ID,
+  Inject
 } from '@angular/core';
 import { trimLabel } from '../common/trim-label.helper';
 import { roundedRect } from '../common/shape.helper';
 import { escapeLabel } from '../common/label.helper';
 import { decimalChecker, count } from '../common/count/count.helper';
 import { GridData } from '../common/grid-layout.helper';
+import { isPlatformBrowser, isPlatformServer } from '@angular/common';
+import { calculateTextWidth } from '../utils/calculate-width';
+import { VERDANA_FONT_WIDTHS_16_PX } from '../common/constants/font-widths';
 
 @Component({
   selector: 'g[ngx-charts-card]',
@@ -103,7 +108,12 @@ export class CardComponent implements OnChanges, OnDestroy {
 
   bandPath: string;
 
-  constructor(element: ElementRef, private cd: ChangeDetectorRef, private zone: NgZone) {
+  constructor(
+    element: ElementRef,
+    private cd: ChangeDetectorRef,
+    private zone: NgZone,
+    @Inject(PLATFORM_ID) private platformId: any
+  ) {
     this.element = element.nativeElement;
   }
 
@@ -111,8 +121,16 @@ export class CardComponent implements OnChanges, OnDestroy {
     this.update();
   }
 
+  ngOnInit() {
+    if (isPlatformServer(this.platformId)) {
+      this.scaleTextSSR();
+    }
+  }
+
   ngOnDestroy(): void {
-    cancelAnimationFrame(this.animationReq);
+    if (isPlatformBrowser(this.platformId)) {
+      cancelAnimationFrame(this.animationReq);
+    }
   }
 
   update(): void {
@@ -146,7 +164,9 @@ export class CardComponent implements OnChanges, OnDestroy {
       this.bandPath = roundedRect(0, 0, this.cardWidth, this.bandHeight, 3, [false, false, true, true]);
 
       setTimeout(() => {
-        this.scaleText();
+        if (isPlatformBrowser(this.platformId)) {
+          this.scaleText();
+        }
         this.value = value;
         if (hasValue && !this.initialized) {
           setTimeout(() => this.startCount(), 20);
@@ -206,7 +226,22 @@ export class CardComponent implements OnChanges, OnDestroy {
     });
   }
 
-  setPadding(): void {
+  scaleTextSSR() {
+    const width = calculateTextWidth(VERDANA_FONT_WIDTHS_16_PX, this.value, 10);
+    const height = 18;
+    const textPadding = (this.textPadding[1] = this.textPadding[3] = this.cardWidth / 8);
+    const availableWidth = this.cardWidth - 2 * textPadding;
+    const availableHeight = this.cardHeight / 3;
+
+    const resizeScale = Math.min(availableWidth / width, availableHeight / height);
+
+    this.textFontSize = Math.floor(this.textFontSize * resizeScale);
+    this.labelFontSize = Math.min(this.textFontSize, 15);
+
+    this.setPadding();
+  }
+
+  setPadding() {
     this.textPadding[1] = this.textPadding[3] = this.cardWidth / 8;
     const padding = this.cardHeight / 2;
     this.textPadding[0] = padding - this.textFontSize - this.labelFontSize / 2;